requirements.txt
pytest~=8.3.3
playwright~=1.47.0
pytest-playwright==0.5.2
allure-pytest==2.13.5